Use original classes for Country deserializing

+++ src/main/java/com/maxmind/geoip2/model/CityResponse.java
@@ -122,7 +122,7 @@ public final class CityResponse extends AbstractCityResponse {
                     model.getTraits().getStaticIpScore()
                 ) :
                 new Traits(
-                    null,
+                    (Integer) null,
                     null,
                     null,
                     null,
